Title: 최솟값 3개, Time: 1027ms, Memory: 25MB, Status: Passed - Codetree

+++ 250626/최솟값 3개/top-3-smallest-number.java
@@ -21,7 +21,7 @@ public class Main {
             int a = small.poll();
             int b = small.poll();
             int c = small.poll();
-            System.out.println(a * b * c);
+            System.out.println((long)a * b * c);
             small.add(a);
             small.add(b);
             small.add(c);
